paint-timing/child-painting-first-image.html

<!DOCTYPE html>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
async_test(function (t) {
    // This test must take place within an iframe to avoid interference from
    // paint entries created by testharness.js
    const parent = document.createElement('iframe');
    document.body.appendChild(parent);

    parent.contentWindow.addEventListener('message', t.step_func(e => {
        assert_equals(e.data, '2 paint first-paint paint first-contentful-paint');
        // When only child frame paints, expect only first-paint.
        t.step_timeout( function() {
            const bufferedEntries = parent.contentWindow.performance.getEntriesByType('paint');
            assert_equals(bufferedEntries.length, 1);
            assert_equals(bufferedEntries[0].entryType, 'paint');
            assert_equals(bufferedEntries[0].name, 'first-paint');
            t.done();
        }, 50);
    }));
    const iframe = parent.contentDocument.createElement('iframe');
    iframe.id = 'child-iframe';
    iframe.src = 'resources/subframe-painting.html';
    parent.contentDocument.body.appendChild(iframe);
}, 'Parent frame ignores paint-timing events fired from child image rendering.');
</script>
</body>
</html>
